backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2
requests>=2.31.0 